Implement new warp list interface with mc 1.7 chat
Using new interface from the minestar library for sending formatted
message with hover and click events via tell raw command.

This allows us to build a simple interface for the user only using the
chat.

The changes in the warp list command are an example for the possiblities
of the API.